\name{NumericalMixingMatrix}
\alias{NumericalMixingMatrix}
\docType{data}
\title{
mixing matrix of data NumericalMixMCF7HS27
}
\description{
real mixing matrix of data NumericalMixMCF7HS27
}
\usage{data(NumericalMixingMatrix)}
\format{
  The format is:
 num [1:2, 1:2] 0.775 0.15 0.225 0.85
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "V1" "V2"
}

\examples{
data(NumericalMixingMatrix)
str(NumericalMixingMatrix) 
}
\keyword{datasets}
